requirements.txt
Django>=5.2.0
psycopg2-binary>=2.9.0
mongoengine>=0.29.0
pymongo>=4.15.0
django-allauth>=65.0.0
cryptography>=46.0.0
Pillow>=11.0.0
requests>=2.32.0
PyJWT>=2.10.0
python-decouple>=3.8
whitenoise>=6.11.0
gunicorn>=23.0.0
dj-database-url>=2.1.0